+++ python/ray/data/_internal/logical/rules/__init__.py
@@ -1,0 +1,3 @@
+from ray.data._internal.logical.rules.operator_fusion import OperatorFusionRule
+
+__all__ = ["OperatorFusionRule"]
